Fixed Maven support with PiTest -> Added forced statics, even, if Lombok's UtilityClass adds them by auto.
