+++ pom.xml
@@ -14,13 +14,13 @@
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <maven.compiler.source>14</maven.compiler.source>
         <maven.compiler.target>14</maven.compiler.target>
-        <junit.version>[5.6.2,)</junit.version>
-        <pitest.version>1.5.1</pitest.version>
-        <surefire.version>3.0.0-M4</surefire.version>
-        <pitest.junit5.version>0.12</pitest.junit5.version>
         <maven.compiler.version>3.8.1</maven.compiler.version>
         <maven.compiler.release>11</maven.compiler.release>
         <lombok.version>1.18.12</lombok.version>
+        <my.junit.version>[5.4.0,)</my.junit.version>
+        <pitest.version>1.5.1</pitest.version>
+        <surefire.version>3.0.0-M4</surefire.version>
+        <pitest.junit5.version>0.12</pitest.junit5.version>
         <org.json.version>20190722</org.json.version>
     </properties>
 
@@ -31,7 +31,6 @@
             <artifactId>json</artifactId>
             <version>${org.json.version}</version>
         </dependency>
-
         <!-- https://mvnrepository.com/artifact/org.projectlombok/lombok -->
         <dependency>
             <groupId>org.projectlombok</groupId>
@@ -39,14 +38,21 @@
             <version>${lombok.version}</version>
             <scope>provided</scope>
         </dependency>
-
+        <!-- https://mvnrepository.com/artifact/org.junit.jupiter/junit-jupiter-api -->
+        <dependency>
+            <groupId>org.junit.jupiter</groupId>
+            <artifactId>junit-jupiter-api</artifactId>
+            <version>${my.junit.version}</version>
+            <scope>test</scope>
+        </dependency>
         <!-- junit 5, unit test -->
         <dependency>
             <groupId>org.junit.jupiter</groupId>
             <artifactId>junit-jupiter-engine</artifactId>
-            <version>${junit.version}</version>
+            <version>${my.junit.version}</version>
             <scope>test</scope>
         </dependency>
+        <!-- IntelliJ IDEA annotatinos -->
         <dependency>
             <groupId>org.jetbrains</groupId>
             <artifactId>annotations</artifactId>
@@ -60,35 +66,68 @@
             <version>${pitest.version}</version>
             <type>pom</type>
         </dependency>-->
+        <!-- https://mvnrepository.com/artifact/org.apache.maven.plugins/maven-failsafe-plugin -->
+        <dependency>
+            <groupId>org.apache.maven.plugins</groupId>
+            <artifactId>maven-failsafe-plugin</artifactId>
+            <version>3.0.0-M4</version>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 
     <build>
+        <sourceDirectory>src</sourceDirectory>
+        <testSourceDirectory>test</testSourceDirectory>
         <plugins>
             <!-- Run as Java 14 -->
             <!-- https://mvnrepository.com/artifact/org.apache.maven.plugins/maven-compiler-plugin -->
             <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
+                <!--<groupId>org.apache.maven.plugins</groupId>-->
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>${maven.compiler.version}</version>
                 <configuration>
                     <release>14</release>
-                    <compilerArgs>--enable-preview</compilerArgs>
+                    <compilerArgs>
+                        --enable-preview
+                    </compilerArgs>
                 </configuration>
             </plugin>
             <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
+                <!--<groupId>org.apache.maven.plugins</groupId>-->
                 <artifactId>maven-surefire-plugin</artifactId>
                 <version>${surefire.version}</version>
                 <configuration>
                     <argLine>--enable-preview</argLine>
                 </configuration>
             </plugin>
+            <plugin>
+                <artifactId>maven-failsafe-plugin</artifactId>
+                <version>${surefire.version}</version>
+                <configuration>
+                    <argLine>--enable-preview</argLine>
+                </configuration>
+            </plugin>
+            <!--<plugin>
+                <artifactId>maven-compiler-plugin</artifactId>
+                <configuration>
+                    <source>${maven.compiler.source}</source>
+                    <target>${maven.compiler.target}</target>
+                </configuration>
+            </plugin>-->
             <!-- Mutation tests with PiTest via JUnit5 -->
             <plugin>
                 <!-- https://mvnrepository.com/artifact/org.pitest/pitest-maven -->
                 <groupId>org.pitest</groupId>
                 <artifactId>pitest-maven</artifactId>
                 <version>${pitest.version}</version>
+
+                <!--<reportSets>
+                    <reportSet>
+                        <reports>
+                            <report>report</report>
+                        </reports>
+                    </reportSet>
+                </reportSets>-->
 
                 <executions>
                     <execution>
@@ -108,6 +147,7 @@
                         <groupId>org.pitest</groupId>
                         <artifactId>pitest-junit5-plugin</artifactId>
                         <version>${pitest.junit5.version}</version>
+
                     </dependency>
                 </dependencies>
                 <configuration>
@@ -115,11 +155,13 @@
                         <param>org.ttrzcinski.utils.*</param>
                     </targetClasses>
                     <targetTests>
-                        <param>test.ttrzcinski.utils.*</param>
+                        <param>org.ttrzcinski.utils.*</param>
                     </targetTests>
+                    <jvmArgs>
+                        --enable-preview
+                    </jvmArgs>
                 </configuration>
             </plugin>
-
         </plugins>
     </build>
